Refactor database interactions by moving db helper to utils and implementing insert/update functions for catalogs and collections
